+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### 🚀 Added
+- Configurable dump transaction (whether to use, an isolation level)
+  [#99](https://github.com/datanymizer/datanymizer/pull/99) ([@evgeniy-r](https://github.com/evgeniy-r))
+
+### ⚙️ Changed
+
+### 🛠 Fixed
+
 ## [v0.3.1] - 2021-09-20
 ### 🚀 Added
 - `ExtData` trait, dictionaries for the Russian locale (person and company names), the new transformers: `middle_name`